+++ Config/routes.php
@@ -54,7 +54,8 @@ Router::connect('/posts/*',			array('controller' => 'posts', 'action' => 'index'
 Router::connect('/post/*',			array('controller' => 'posts', 'action' => 'view',		'plugin' => 'me_cms'));
 
 //Profiles controller
-Router::connect('/newpassword/request', array('controller' => 'profiles', 'action' => 'request_new_password', 'plugin' => 'me_cms'));
+Router::connect('/password/request',		array('controller' => 'profiles', 'action' => 'request_new_password',	'plugin' => 'me_cms'));
+Router::connect('/password/:id/:token',	array('controller' => 'profiles', 'action' => 'reset_password',			'plugin' => 'me_cms'), array('pass' => array('id', 'token')));
 
 //System controller
 Router::connect('/offline', array('controller' => 'systems', 'action' => 'offline', 'plugin' => 'me_cms'));